- Clan priority done (not stored in DB yet).

+++ TourismReco2/Models/Clan.cs
@@ -6,7 +6,7 @@ namespace TourismReco2.Models
     {
         public int ClanId { get; set; }
         public string ClanName { get; set; }
-        public float? ClanPreference { get; set; }
+        public double? ClanPreference { get; set; }
 
         public bool IsSelected { get; set; }
